# tests/contract/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "EnvelopeModel_QuarantineDetail_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/QuarantineDetail"
          },
          "links": {
            "$ref": "#/components/schemas/LinksModel"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaModel"
          }
        },
        "required": [
          "data",
          "meta",
          "links"
        ],
        "title": "EnvelopeModel[QuarantineDetail]",
        "type": "object"
      },
      "EnvelopeModel_QuarantineSummary_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/QuarantineSummary"
          },
          "links": {
            "$ref": "#/components/schemas/LinksModel"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaModel"
          }
        },
        "required": [
          "data",
          "meta",
          "links"
        ],
        "title": "EnvelopeModel[QuarantineSummary]",
        "type": "object"
      },
...
      "EnvelopeModel_list_QuarantineRow__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/QuarantineRow"
            },
            "title": "Data",
            "type": "array"
          },
          "links": {
            "$ref": "#/components/schemas/LinksModel"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaModel"
          }
        },
        "required": [
          "data",
          "meta",
          "links"
        ],
        "title": "EnvelopeModel[list[QuarantineRow]]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LinksModel": {
        "additionalProperties": true,
        "properties": {
          "explain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pre-built /v1/explain call for every handle in the body.",
            "title": "Explain"
          },
          "next": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Next page, or null at the end.",
            "title": "Next"
          },
          "self": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Canonical path of this response.",
            "title": "Self"
          }
        },
        "title": "LinksModel",
        "type": "object"
      },
...
      "Problem": {
        "additionalProperties": true,
        "description": "The wire shape of every failure (RFC 9457).",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "What went wrong on this request.",
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-field failures, each with a JSON Pointer.",
            "title": "Errors"
          },
          "instance": {
            "description": "Path of the request that failed.",
            "title": "Instance",
            "type": "string"
          },
          "request_id": {
            "description": "ULID joining this response to the audit trail.",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code, repeated in the body per RFC 9457.",
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short, human-readable summary of the error code.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "Absolute URI of the error code; resolves at /v1/errors/{code}.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "instance",
          "request_id"
        ],
        "title": "Problem",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "description": "Static owner key for this deployment. /healthz and the frontend are open.",
        "in": "header",
        "name": "X-Glasswell-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Glass-box upstream analytics on public data. Every number this API serves carries a\nderivation handle in band, and `GET /v1/explain` walks that handle back to the\nchecksummed government file it came from.\n\nConventions that hold everywhere: responses are `data` / `meta` / `links`; collections\nare cursor-paginated with no offset parameter; failures are RFC 9457 problem documents\nwhose `type` resolves at `/v1/errors/{code}`; and `as_of` selects knowledge time, with\nthe resolved vintage reported back in `meta.as_of`.\n\nThis deployment is the North Dakota slice: wells, geometry, monthly production, the\nlineage spine, the conformance registry, quarantine and the glossary. Forecasts,\neconomics and other basins are not served.",
    "title": "glasswell",
    "version": "0.1.0",
    "x-glasswell-freeze": {
      "checked_by": "tests/contract/openapi_diff.py",
      "criterion": "S1",
      "frozen_on": "2026-08-20",
      "policy": "Additive change only. A path, operation, parameter, response field or enum value that is published here is published for the life of /v1; removing one, or making an optional request parameter required, is a /v2 event (blueprint \u00a73.6.1).",
      "status": "frozen",
      "surface": "v1"
    }
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/quarantine": {
      "get": {
        "description": "Rows a conformance rule rejected, with the reason, the rule, how many fetches have re-presented them and their lifecycle state. Nothing is dropped silently: an empty quarantine means the rules are not firing, not that the data is clean.",
        "operationId": "list_quarantine",
        "parameters": [
          {
            "description": "Opaque cursor from meta.next_cursor. There is no offset parameter.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque cursor from meta.next_cursor. There is no offset parameter.",
              "title": "Cursor"
            }
          },
          {
            "description": "Page size, 100 by default, 200 at most (SB-07 \u00a79.5 caps the spine collections lower than the data ones).",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "description": "Page size, 100 by default, 200 at most (SB-07 \u00a79.5 caps the spine collections lower than the data ones).",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Filter to one source.",
            "in": "query",
            "name": "source_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one source.",
              "title": "Source Id"
            }
          },
          {
            "description": "Filter to one reason code.",
            "in": "query",
            "name": "reason_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one reason code.",
              "title": "Reason Code"
            }
          },
          {
            "description": "Filter to one rejecting rule.",
            "in": "query",
            "name": "rule_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one rejecting rule.",
              "title": "Rule Id"
            }
          },
          {
            "description": "Filter to one lifecycle state.",
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "open",
                    "released",
                    "accepted_loss",
                    "superseded"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one lifecycle state.",
              "title": "State"
            }
          },
          {
            "description": "Filter to one pipeline stage.",
            "in": "query",
            "name": "stage",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "parse",
                    "validate",
                    "conform",
                    "join"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one pipeline stage.",
              "title": "Stage"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeModel_list_QuarantineRow__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "examples": {
                  "cursor_malformed": {
                    "summary": "Cursor malformed",
                    "value": {
                      "detail": "The cursor did not decode to the four declared fields.",
                      "instance": "/v1",
                      "request_id": "01JBQ7M0Z8K2V4N6X8R0T2Y4W6",
                      "status": 422,
                      "title": "Cursor malformed",
                      "type": "/v1/errors/cursor_malformed"
                    }
                  },
                  "cursor_query_mismatch": {
                    "summary": "Cursor does not match this query",
                    "value": {
                      "detail": "The cursor was minted against a different filter set. Continuing would return a page from a different result set (SB-04 \u00a72.3).",
                      "instance": "/v1",
                      "request_id": "01JBQ7M0Z8K2V4N6X8R0T2Y4W6",
                      "status": 422,
                      "title": "Cursor does not match this query",
                      "type": "/v1/errors/cursor_query_mismatch"
                    }
                  },
                  "validation_failed": {
                    "summary": "Request validation failed",
                    "value": {
                      "detail": "A parameter failed validation, including declared caps. `errors[]` names each offending pointer. Caps are refused, never silently clamped.",
                      "instance": "/v1",
                      "request_id": "01JBQ7M0Z8K2V4N6X8R0T2Y4W6",
                      "status": 422,
                      "title": "Request validation failed",
                      "type": "/v1/errors/validation_failed"
                    }
                  }
                }
              }
            },
            "description": "Request validation failed"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "examples": {
                  "service_degraded": {
                    "summary": "Service degraded",
                    "value": {
                      "detail": "A required store is unavailable.",
                      "instance": "/v1",
                      "request_id": "01JBQ7M0Z8K2V4N6X8R0T2Y4W6",
                      "status": 503,
                      "title": "Service degraded",
                      "type": "/v1/errors/service_degraded"
                    }
                  }
                }
              }
            },
            "description": "Service degraded"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List quarantined rows",
        "tags": [
          "quality"
        ],
        "x-glasswell-dataset": {
          "anchors": [],
          "collection_pointer": "",
          "columns": {
            "default": [
              "/quarantine_id",
              "/reason_code",
              "/rule_id",
              "/state",
              "/stage",
              "/occurrence_count",
              "/last_seen_at"
            ],
            "hidden": [
              "/row_fingerprint",
              "/notes"
            ],
            "hidden_reason": {
              "/notes": "Free text a reviewer left on release; empty on almost every row and not a dimension anything is filtered by.",
              "/row_fingerprint": "A content address over the rejected row's bytes, useful for joining two fetches of the same row and unreadable as a column."
            },
            "sort": "/last_seen_at"
          },
          "detail_operation": "get_quarantine_row",
          "facets": [
            "source_id",
            "reason_code",
            "rule_id",
            "state",
            "stage"
          ],
          "group": "kitchen",
          "id": "quarantine",
          "intro": "nb_dataset_quarantine",
          "order": 20,
          "row_id": [
            "/quarantine_id"
          ],
          "summary_operation": "get_quarantine_summary",
          "title": "Quarantine"
        },
        "x-glasswell-request-example": {
          "path": {},
          "query": {
            "limit": 5
          }
        }
      }
    },
    "/v1/quarantine/summary": {
      "get": {
        "description": "Counts and shares by reason code or by stage \u2014 the scorecard view of what the rules are rejecting. Declared before /quarantine/{quarantine_id} so `summary` is not matched as an id.",
        "operationId": "get_quarantine_summary",
        "parameters": [
          {
            "description": "Dimension to group by.",
            "in": "query",
            "name": "group_by",
            "required": false,
            "schema": {
              "default": "reason_code",
              "description": "Dimension to group by.",
              "enum": [
                "reason_code",
                "stage"
              ],
              "title": "Group By",
              "type": "string"
            }
          },
          {
            "description": "Filter to one source.",
            "in": "query",
            "name": "source_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one source.",
              "title": "Source Id"
            }
          },
          {
            "description": "Filter to one lifecycle state.",
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one lifecycle state.",
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeModel_QuarantineSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "examples": {
                  "validation_failed": {
                    "summary": "Request validation failed",
                    "value": {
                      "detail": "A parameter failed validation, including declared caps. `errors[]` names each offending pointer. Caps are refused, never silently clamped.",
                      "instance": "/v1",
                      "request_id": "01JBQ7M0Z8K2V4N6X8R0T2Y4W6",
                      "status": 422,
                      "title": "Request validation failed",
                      "type": "/v1/errors/validation_failed"
                    }
                  }
                }
              }
            },
            "description": "Request validation failed"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "examples": {
                  "service_degraded": {
                    "summary": "Service degraded",
                    "value": {
                      "detail": "A required store is unavailable.",
                      "instance": "/v1",
                      "request_id": "01JBQ7M0Z8K2V4N6X8R0T2Y4W6",
                      "status": 503,
                      "title": "Service degraded",
                      "type": "/v1/errors/service_degraded"
                    }
                  }
                }
              }
            },
            "description": "Service degraded"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Quarantine shares",
        "tags": [
          "quality"
        ],
        "x-glasswell-request-example": {
          "path": {},
          "query": {
            "group_by": "reason_code"
          }
        }
      }
    },
    "/v1/quarantine/{quarantine_id}": {
      "get": {
        "description": "The rejected row itself, served verbatim alongside the rule that rejected it and the manifests it was first and last seen in. Re-processing always re-reads the manifest, never this payload, so the payload is evidence rather than input. The example id is the contract fixture's. `drv_`, `man_` and `qr_` ids are content addresses over bytes and run parameters, so they differ per deployment by construction \u2014 take a live one from the `d` handle on any served figure, or from `/v1/manifests`.",
        "operationId": "get_quarantine_row",
        "parameters": [
          {
            "description": "Quarantined row id.",
            "in": "path",
            "name": "quarantine_id",
            "required": true,
            "schema": {
              "description": "Quarantined row id.",
              "title": "Quarantine Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeModel_QuarantineDetail_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "examples": {
                  "not_found": {
                    "summary": "Not found",
                    "value": {
                      "detail": "The resource does not exist, or exists but is not visible to this principal. Both answer the same way so probing reveals nothing.",
                      "instance": "/v1",
                      "request_id": "01JBQ7M0Z8K2V4N6X8R0T2Y4W6",
                      "status": 404,
                      "title": "Not found",
                      "type": "/v1/errors/not_found"
                    }
                  }
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "examples": {
                  "service_degraded": {
                    "summary": "Service degraded",
                    "value": {
                      "detail": "A required store is unavailable.",
                      "instance": "/v1",
                      "request_id": "01JBQ7M0Z8K2V4N6X8R0T2Y4W6",
                      "status": 503,
                      "title": "Service degraded",
                      "type": "/v1/errors/service_degraded"
                    }
                  }
                }
              }
            },
            "description": "Service degraded"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "One quarantined row",
        "tags": [
          "quality"
        ],
        "x-glasswell-request-example": {
          "path": {
            "quarantine_id": "qr_01contract0001"
          },
          "query": {}
        }
      }
    },
...
  }
}